requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
